alternative/requirements.txt
Pillow
pytesseract